Update SSSet.js
Rewrote the storage component to use object properties instead of a binary-searched array. Things are quite a lot faster now since it can rely on the built-in hasOwnProperty method to check for existence and there is no need to explicitly sort on insert. The hash function used to generate object keys is currently JSON.stringify() but could be changed to something better/faster.